refactored parse_focused_problems to pull out tsv parsing with pickles
The parse_focused_problems_via_api no longer contains the logic
for parsing the input tsv file.  This should be shared code with
the _via_py_umls function as well. This new parse_focused_prob_tsv
function will be run if there isn't already a pickle of the parsed
tsv data structures.  When run (and a partials_directory provided),
a new pickle file will be created so we don't have to re-process
it every time.