check for partials file for a cui before processing; partials 4 allergens
If we're running in order, then if a pkl file exists for a given
head_cui, it has already been processed and we can skip it and
go directly to the next head_cui to process.

Sorting the problem cuis before processing helps make the order
consistent, which helps with reloading partials reliably.

added partials support for allergens. This hasn't been as thoroughly
refactored as the problems pipeline because the allergens file
parsing and concept extraction is more interwoven.