global vars; added medications and labs
Refactored some of the variables to be global rather than local
to a function (like the node_map entries and column names for the
various RRF files).

--source-type was added as a command line option but won't actually
be useful for meds but helps with labs.

The general medication algorithm is to:
1) walk through the ATC1-4 top level (meaning first two levels)
2) write those to both INGREDIENTS and BRANDS output ttl files
3) walk through for all RxNorm + BN entries
4) walk through the relations file looking for tradename relations
   that link back to our brandnames (IN/MIN <-> BN)
6) walk through the ATC1-4 listing for any ingredients (IN/MIN)
7) write those to the ingredients list
8) write those as brands with the ingredient's ATC1-4 code as parent

The general outline for labs is much simpler
1) find all CUIs that fall under the provided semantic types
2) walk the MRCONSO file looking for those CUIs and make sure
   they're no suppressed, are preferred, etc. and match the
   --source-type provided
3) write these to the ttl

Of course, we're skipping all suppressed entries for all concepts
above.